espire alpha 0.1

+++ mainapp/views.py
@@ -1,9 +1,9 @@
 from django.shortcuts import render, HttpResponse, HttpResponseRedirect
-from mainapp.models import Brand, Product, Cart
+from mainapp.models import Brand, Product, Cart, Favourite
 from userapp.forms import UserLoginForm
 from userapp.models import User
 from django.contrib import auth
-from userapp.views import get_cart
+from userapp.views import get_cart, get_favourite
 import uuid
 # Create your views here.
 
@@ -20,11 +20,18 @@ def goods(request, context=None):
     print(request.session['device_id'])
     if request.method == 'GET':
         if not context:
+            favourites = get_favourite(request)
+            id_favourites = [x.product_id for x in favourites]
+            
             context = {'goods':Product.objects.all(),
                        'brands':Brand.objects.all(),
                        'form':UserLoginForm, 
-                       'carts':get_cart(request), 
-                       'range':range(len(get_cart(request)))}
+                       'carts':get_cart(request),
+                       'favourites':favourites, 
+                       'id_favourites':id_favourites, 
+                       'range':range(len(get_cart(request))),
+                       'range_goods': range(len(Product.objects.all()))}
+            print(context['favourites'],context['goods'])
         return render(request,'mainapp/goods.html', context)
     else:
         form = UserLoginForm(data=request.POST)
@@ -45,14 +52,15 @@ def compare_carts(user,device_id):
     carts2 = Cart.objects.filter(user=user)
     sorted_carts = carts2.order_by('product')
     i = 1
-    last_product = sorted_carts[0].product
-    while i < len(sorted_carts):
-        if last_product == sorted_carts[i].product:
-            sorted_carts[i].delete()
-        else:
-            last_product = sorted_carts[i].product
-        i += 1
-        print(sorted_carts)
+    if len(sorted_carts) > 0:
+        last_product = sorted_carts[0].product
+        while i < len(sorted_carts):
+            if last_product == sorted_carts[i].product:
+                sorted_carts[i].delete()
+            else:
+                last_product = sorted_carts[i].product
+            i += 1
+            print(sorted_carts)
 
 def good(request):
     try:
